pianoroll: amend 5ac3013cbc (type mismatch)

+++ gtk2_ardour/midi_view_background.cc
@@ -264,7 +264,7 @@ MidiViewBackground::apply_note_range (uint8_t lowest, uint8_t highest, bool to_c
 	/* Enforce a 1 octave minimum */
 
 	if (highest - lowest < 11) {
-		int8_t mid = lowest + ((highest - lowest) / 2);
+		int mid = lowest + ((highest - lowest) / 2);
 		mid = max(6, min(121, mid));
 		lowest = mid - 6;
 		highest = lowest + 11;
